requirements.txt
absl-py
aiohttp
anyio
argon2-cffi
async-generator
async-timeout==3.0.1
attrs
autopep8==1.5.7
Babel
backcall
bleach
blinker==1.4
brotlipy==0.7.0
cachetools
certifi==2021.5.30
cffi
chardet
click
colorama
cryptography
cycler==0.10.0
decorator
defusedxml
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
gdown==3.13.0
google-auth
google-auth-oauthlib==0.4.1
grpcio
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
itsdangerous==2.0.1
jedi==0.17.0
Jinja2 
json5
jsonschema
jupyter-client
jupyter-core
jupyter-packaging 
jupyter-server 
jupyterlab 
jupyterlab-pygments 
jupyterlab-server 
kiwisolver==1.3.1
Markdown 
MarkupSafe 
matplotlib==3.4.2
mistune==0.8.4
multidict
nbclassic 
nbclient 
nbconvert 
nbformat
nest-asyncio 
notebook 
numpy
oauthlib
opencv-python==4.5.3.56
packaging 
pandas==1.3.1
pandocfilters
parso 
pickleshare 
Pillow==8.3.1
prometheus-client
prompt-toolkit
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.7.0
pycparser 
Pygments 
PyJWT 
PyMySQL==1.0.2
pyOpenSSL 
pyparsing 
pyrsistent 
PySocks
python-dateutil 
pytz
pyu2f 
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
requests
requests-oauthlib
rsa
scipy==1.7.1
seaborn==0.11.1
Send2Trash
six
sniffio
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
terminado==0.9.4
testpath
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado
tqdm==4.62.0
traitlets
typing-extensions
urllib3
watchdog==2.1.3
wcwidth
webencodings==0.5.1
Werkzeug
win-inet-pton
wincertstore==0.2
yarl
zipp

